Rename _encoder to encoder under _components module
