+++ fastkafka/_components/encoder/avro.py
