+++ fastkafka/_components/encoder/json.py
