+++ fastkafka/_modidx.py
@@ -93,26 +93,6 @@ d = { 'settings': { 'branch': 'main',
                                                                                                   'fastkafka/_application/tester.py'),
                                                'fastkafka._application.tester.mirror_producer': ( 'tester.html#mirror_producer',
                                                                                                   'fastkafka/_application/tester.py')},
-            'fastkafka._components._encoder.avro': { 'fastkafka._components._encoder.avro.AvroBase': ( 'avro_encode_decoder.html#avrobase',
-                                                                                                       'fastkafka/_components/_encoder/avro.py'),
-                                                     'fastkafka._components._encoder.avro.AvroBase._avro_schema': ( 'avro_encode_decoder.html#avrobase._avro_schema',
-                                                                                                                    'fastkafka/_components/_encoder/avro.py'),
-                                                     'fastkafka._components._encoder.avro.AvroBase.avro_schema': ( 'avro_encode_decoder.html#avrobase.avro_schema',
-                                                                                                                   'fastkafka/_components/_encoder/avro.py'),
-                                                     'fastkafka._components._encoder.avro.AvroBase.avro_schema_for_pydantic': ( 'avro_encode_decoder.html#avrobase.avro_schema_for_pydantic',
-                                                                                                                                'fastkafka/_components/_encoder/avro.py'),
-                                                     'fastkafka._components._encoder.avro.avro_decoder': ( 'avro_encode_decoder.html#avro_decoder',
-                                                                                                           'fastkafka/_components/_encoder/avro.py'),
-                                                     'fastkafka._components._encoder.avro.avro_encoder': ( 'avro_encode_decoder.html#avro_encoder',
-                                                                                                           'fastkafka/_components/_encoder/avro.py'),
-                                                     'fastkafka._components._encoder.avro.avsc_to_pydantic': ( 'avro_encode_decoder.html#avsc_to_pydantic',
-                                                                                                               'fastkafka/_components/_encoder/avro.py')},
-            'fastkafka._components._encoder.json': { 'fastkafka._components._encoder.json._to_json_utf8': ( 'json_encode_decoder.html#_to_json_utf8',
-                                                                                                            'fastkafka/_components/_encoder/json.py'),
-                                                     'fastkafka._components._encoder.json.json_decoder': ( 'json_encode_decoder.html#json_decoder',
-                                                                                                           'fastkafka/_components/_encoder/json.py'),
-                                                     'fastkafka._components._encoder.json.json_encoder': ( 'json_encode_decoder.html#json_encoder',
-                                                                                                           'fastkafka/_components/_encoder/json.py')},
             'fastkafka._components.aiokafka_consumer_loop': { 'fastkafka._components.aiokafka_consumer_loop._aiokafka_consumer_loop': ( 'consumerloop.html#_aiokafka_consumer_loop',
                                                                                                                                         'fastkafka/_components/aiokafka_consumer_loop.py'),
                                                               'fastkafka._components.aiokafka_consumer_loop._create_safe_callback': ( 'consumerloop.html#_create_safe_callback',
@@ -201,6 +181,26 @@ d = { 'settings': { 'branch': 'main',
                                                                                                                              'fastkafka/_components/docs_dependencies.py'),
                                                          'fastkafka._components.docs_dependencies._install_node': ( 'docs_dependencies.html#_install_node',
                                                                                                                     'fastkafka/_components/docs_dependencies.py')},
+            'fastkafka._components.encoder.avro': { 'fastkafka._components.encoder.avro.AvroBase': ( 'avro_encode_decoder.html#avrobase',
+                                                                                                     'fastkafka/_components/encoder/avro.py'),
+                                                    'fastkafka._components.encoder.avro.AvroBase._avro_schema': ( 'avro_encode_decoder.html#avrobase._avro_schema',
+                                                                                                                  'fastkafka/_components/encoder/avro.py'),
+                                                    'fastkafka._components.encoder.avro.AvroBase.avro_schema': ( 'avro_encode_decoder.html#avrobase.avro_schema',
+                                                                                                                 'fastkafka/_components/encoder/avro.py'),
+                                                    'fastkafka._components.encoder.avro.AvroBase.avro_schema_for_pydantic': ( 'avro_encode_decoder.html#avrobase.avro_schema_for_pydantic',
+                                                                                                                              'fastkafka/_components/encoder/avro.py'),
+                                                    'fastkafka._components.encoder.avro.avro_decoder': ( 'avro_encode_decoder.html#avro_decoder',
+                                                                                                         'fastkafka/_components/encoder/avro.py'),
+                                                    'fastkafka._components.encoder.avro.avro_encoder': ( 'avro_encode_decoder.html#avro_encoder',
+                                                                                                         'fastkafka/_components/encoder/avro.py'),
+                                                    'fastkafka._components.encoder.avro.avsc_to_pydantic': ( 'avro_encode_decoder.html#avsc_to_pydantic',
+                                                                                                             'fastkafka/_components/encoder/avro.py')},
+            'fastkafka._components.encoder.json': { 'fastkafka._components.encoder.json._to_json_utf8': ( 'json_encode_decoder.html#_to_json_utf8',
+                                                                                                          'fastkafka/_components/encoder/json.py'),
+                                                    'fastkafka._components.encoder.json.json_decoder': ( 'json_encode_decoder.html#json_decoder',
+                                                                                                         'fastkafka/_components/encoder/json.py'),
+                                                    'fastkafka._components.encoder.json.json_encoder': ( 'json_encode_decoder.html#json_encoder',
+                                                                                                         'fastkafka/_components/encoder/json.py')},
             'fastkafka._components.helpers': { 'fastkafka._components.helpers.ImportFromStringError': ( 'internal_helpers.html#importfromstringerror',
                                                                                                         'fastkafka/_components/helpers.py'),
                                                'fastkafka._components.helpers._import_from_string': ( 'internal_helpers.html#_import_from_string',
